fulltext, msg_engine, windowtext を分離した。
